refactor: guarantee non-null decimals + TokenQueryService interface/impl split
Main change: TokenRegistryCurrencyData.decimals is now guaranteed non-null
for every token the query service returns. TokenQueryService.mergeMetadata
seeds the builder with decimals=0 before applying CIP-26 and CIP-68 overrides,
so tokens that advertise no decimal count in either standard get the canonical
default instead of null. The @Nullable annotation on the field is replaced by
@Nonnull (picked up by Lombok's AllArgsConstructor which now enforces the
contract at build time).

Downstream consequences:
- AccountMapperUtil and TransactionMapperUtils both had a duplicated
  getDecimalsWithFallback() helper applying the same .orElse(0) default.
  Both deleted; callers inline metadata.getDecimals() directly since the
  contract now guarantees non-null.

Interface/impl split:
- TokenQueryService is now an interface in the common.service package
- The implementation moves to TokenQueryServiceImpl (fixes the earlier
  typo'd TokenQueryServiceIImpl class name across 5 files)
- @Override added on queryMetadataBatch in the impl
- TokenQueryService interface gets a Javadoc describing the contract
  (merge priority, CIP-68 prefix conversion, non-null decimals guarantee)

Defensive numeric conversion:
- applyCip26 / applyCip68 use Math.toIntExact(Long) instead of .intValue()
  when narrowing the decimals column. Silent truncation at Integer.MAX_VALUE
  is now a loud IllegalStateException — a nit that will never fire in
  practice (decimal counts are 0-18) but removes a footgun.

Tests: 78/78 pass. TokenQueryServiceTest.shouldLeaveNullDecimalsAsNull
renamed to shouldDefaultDecimalsToZeroWhenBothStandardsLackIt and asserts
equal-to-0 instead of null. TokenRegistryMapperTest.shouldHandleAllNullFields
renamed to shouldHandleAllNullableFieldsNull and seeds decimals=0 on the
builder. Several DataMapperTest and TokenRegistryMapperTest cases needed
.decimals(0) added since the builder now rejects null.
