+++ api/src/main/java/org/cardanofoundation/rosetta/api/account/mapper/AccountMapperUtil.java
@@ -14,7 +14,6 @@ import org.openapitools.client.model.*;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.Nullable;
-import javax.validation.constraints.NotNull;
 import java.math.BigInteger;
 import java.util.*;
 
@@ -64,7 +63,7 @@ public class AccountMapperUtil {
             amounts.add(
                     dataMapper.mapAmount(b.quantity().toString(),
                             symbol,
-                            getDecimalsWithFallback(metadata),
+                            metadata.getDecimals(),
                             metadata)
             );
         }
@@ -125,7 +124,7 @@ public class AccountMapperUtil {
                     Amount tokenAmount = dataMapper.mapAmount(
                             amount.getQuantity().toString(),
                             symbol,
-                            getDecimalsWithFallback(metadata),
+                            metadata.getDecimals(),
                             metadata
                     );
 
@@ -138,11 +137,6 @@ public class AccountMapperUtil {
                 .toList();
 
         return coinTokens.isEmpty() ? null : Map.of(coinIdentifier, coinTokens);
-    }
-
-    private static int getDecimalsWithFallback(@NotNull TokenRegistryCurrencyData metadata) {
-        return Optional.ofNullable(metadata.getDecimals())
-                .orElse(0);
     }
 
     private CurrencyResponse getAdaCurrency() {
