+++ api/src/main/java/org/cardanofoundation/rosetta/api/block/mapper/TransactionMapperUtils.java
@@ -22,7 +22,6 @@ import org.openapitools.client.model.*;
 import org.springframework.stereotype.Component;
 
 import javax.annotation.Nullable;
-import javax.validation.constraints.NotNull;
 import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
@@ -227,14 +226,9 @@ public class TransactionMapperUtils {
     return dataMapper.mapAmount(
             dataMapper.mapValue(amount.getQuantity().toString(), spent),
             symbol,
-            getDecimalsWithFallback(metadata),
+            metadata.getDecimals(),
             metadata
     );
-  }
-
-  private static int getDecimalsWithFallback(@NotNull TokenRegistryCurrencyData metadata) {
-    return Optional.ofNullable(metadata.getDecimals())
-            .orElse(0);
   }
 
 }
