+++ api/src/test/java/org/cardanofoundation/rosetta/api/common/service/TokenRegistryServiceImplTest.java
@@ -28,7 +28,7 @@ import static org.mockito.Mockito.*;
 class TokenRegistryServiceImplTest {
 
     @Mock
-    private TokenQueryService tokenQueryService;
+    private TokenQueryServiceImpl tokenQueryService;
 
     private TokenRegistryServiceImpl tokenRegistryService;
